LayoutTests/fast/scrolling/ios/scroll-events-default-expected.txt

PASS receivedScrollEvent is true
PASS successfullyParsed is true

TEST COMPLETE
